Changed Logic For CleanCartitems

+++ WebApplication1/Controllers/HomeController.cs
@@ -68,7 +68,6 @@ namespace WebApplication1.Controllers
                 }
                 var cartItem = new CartItems
                 {
-
                     ItemID = item.ItemId,
                     ItemName = item.ItemName,
                     Price = 1000M,
@@ -104,8 +103,12 @@ namespace WebApplication1.Controllers
             var expirationDate = DateTime.UtcNow.AddDays(-7);
             var oldCarts = _db.CartItems.Where(c => c.Date < expirationDate);
 
-            _db.CartItems.RemoveRange(oldCarts);
-            _db.SaveChanges();
+            
+            if (oldCarts.Any())
+            {
+                _db.CartItems.RemoveRange(oldCarts);
+                _db.SaveChanges();
+            }
         }
         [HttpGet]
         public IActionResult Clientreg()
